refactor:fix AI comments
