+++ src/utils/users.js
@@ -1,7 +1,7 @@
 const usersServices = require('../services/users/users.services');
 
 const isUserExist = (userId) => {
-  return !!(userId && !isNaN(userId) && usersServices.getUserById(userId));
+  return !!(!isNaN(userId) && usersServices.getUserById(userId));
 };
 
 module.exports = { isUserExist };
